+++ src/mods/scala/domains/marine/Consumer.scala
@@ -7,7 +7,7 @@ import org.apache.pekko.actor.typed.{Behavior, Signal, TypedActorContext}
 import org.apache.pekko.actor.typed.scaladsl.Behaviors
 import org.evrete.api.Knowledge
 
-trait Consumer extends Actor[draco.format.json.JSON]
+trait Consumer extends DracoType
 
 object Consumer extends App with DracoType {
   override lazy val typeDefinition: TypeDefinition = TypeLoader.loadType(TypeName ("Consumer", _namePackage = Seq ("domains", "marine")))